Document general library usage
One of the most common issues people are having with the SDK seems to be that they don't know how to use Vita runtime libraries. This change adds instructions for doing that.

While this probably belongs somewhere in vitasdk eventually, currently this page is kind of the closest thing to documentation the SDK has, so I believe putting this here is a good holdover until someone decides to write a full one.

I tried to be very explicit with the instructions because chances are people who need these instructions are beginners who've never had to touch a build process, start off utterly confused and need at least a little handholding.